separate projects for monix/jdbc/cassandra
