+++ quill-monix-macro/src/main/scala/pl/jozwik/quillgeneric/quillmacro/monix/MonixWithContextDateQuotes.scala
@@ -1,0 +1,7 @@
+package pl.jozwik.quillgeneric.quillmacro.monix
+
+import pl.jozwik.quillgeneric.quillmacro.DateQuotes
+
+object MonixWithContextDateQuotes {
+  type MonixWithContextDateQuotes = MonixWithContext with DateQuotes
+}
